KMK-geo-api/requirements.txt
certifi==2020.6.20
chardet==3.0.4
gtfs-realtime-bindings==0.0.6
idna==2.10
paho-mqtt==1.5.0
protobuf==3.13.0
requests==2.24.0
six==1.15.0
urllib3==1.25.10
